fix: move useMovies hook to correct folder, convert AddMovie/NavBar to async DB-backed movie fetching

+++ src/main/java/com/Springboot/Ticket_Booking_System/repository/MovieRepository.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.Springboot.Ticket_Booking_System.model.Movie;
 
 @Repository
-public interface MovieRepository extends JpaRepository<Movie, Integer> {
-
+public interface MovieRepository extends JpaRepository<Movie, Long> {
+    // No custom queries needed yet — JpaRepository already provides
+    // findAll(), save(), findById(), deleteById(), existsById()
+    // which is everything MovieService currently calls.
 }
